rename abstract classes, add knight move classes
