+++ src/main/java/chessboard/services/transformers/AbstractKnightMoveTransformer.java
@@ -1,0 +1,9 @@
+package chessboard.services.transformers;
+
+/**
+ * Knight move transformer package.
+ *
+ * @author David Hatton
+ */
+public abstract class AbstractKnightMoveTransformer {
+}
