+++ src/test/java/chessboard/services/transformers/AbstractKnightMoveTransformerTest.java
@@ -1,0 +1,17 @@
+package chessboard.services.transformers;
+
+import org.junit.Before;
+
+import static org.junit.Assert.*;
+
+/**
+ * Knight move tests.
+ *
+ * @author David Hatton
+ */
+public class AbstractKnightMoveTransformerTest {
+
+	@Before
+	public void setUp() throws Exception {
+	}
+}
