+++ src/test/java/chessboard/services/transformers/AbstractLateralTransformerTest.java
@@ -10,7 +10,7 @@ import chessboard.exceptions.OutOfBoundsException;
 import chessboard.exceptions.WrongDirectionException;
 import chessboard.services.helpers.Coordinate;
 
-import static chessboard.services.transformers.Lateral.lateral;
+import static chessboard.services.transformers.AbstractLateralTransformer.lateral;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
@@ -20,7 +20,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
  * @author David Hatton
  */
 @RunWith(BlockJUnit4ClassRunner.class)
-public class LateralTest {
+public class AbstractLateralTransformerTest {
 
 	private Coordinate originCoordinate;
